requirements.txt
Flask==2.3.2
flask-restx==1.1.0
werkzeug==2.3.7
mysql-connector-python==8.0.33
python-dotenv
flask-login
